refactor keymap generation to `lib/keyseq`
